fix(harnesses): native-integrations — keep the slice under the repo line caps
Repo-shape CI caught three files over the cap after the Claude in Chrome
commit: launch_extras.rs (658) — its tests move to launch_extras_tests.rs
on the bundles_tests.rs pattern; probe.rs (610 vs its 609 ratchet) — the
catalog probe passes an empty slice and the already-imported BTreeMap,
dropping a dead local and import, ratchet tightened to 608;
manager/tests.rs (601) — two single-level imports merged.
