+++ anyharness/crates/anyharness-lib/src/live/sessions/manager/tests.rs
@@ -16,12 +16,11 @@ use crate::domains::agents::registry::built_in_registry;
 use crate::domains::sessions::model::{SessionEventRecord, SessionRecord};
 use crate::domains::sessions::runtime_event::RuntimeInjectedSessionEvent;
 use crate::domains::sessions::store::SessionStore;
-use crate::live::sessions::actor::command::SessionCommand;
-use crate::live::sessions::handle::LiveSessionHandle;
 use crate::live::sessions::model::{
     LaunchEnv, SessionHooks, SessionLaunch, SessionStartupStrategy, SystemPromptAppends,
 };
 use crate::live::sessions::runtime_events::into_session_event;
+use crate::live::sessions::{actor::command::SessionCommand, handle::LiveSessionHandle};
 use crate::persistence::Db;
 mod startup_ownership;
